Refactor:#76 - standardize select link upload icon naming, location, and format
- Relocate from public/assets/icons/selectLinkUploadIcon.svg to src/assets/icons/select-link-upload.svg
- Apply consistent SVGR props
